Adicionadas novas exceptions personalizadas e criados testes de unidade

+++ src/main/java/com/lambdateam/mycar/service/MaintenanceTypeService.java
@@ -6,38 +6,67 @@ import com.lambdateam.mycar.repository.MaintenanceTypeRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import com.lambdateam.mycar.exception.ExpiredJwtException;
+
 @AllArgsConstructor
 @Service
 public class MaintenanceTypeService {
 
     private final MaintenanceTypeRepository repository;
 
-    private MaintenanceTypeModel findOrThrow(final Long id) {
-        return repository
-                .findById(id)
-                .orElseThrow(
-                        () -> new NotFoundException("The maintenance type id " + id +" was not found.")
-                );
+    private MaintenanceTypeModel findOrThrow(final Long id) throws ExpiredJwtException {
+        try{
+            return repository
+                    .findById(id)
+                    .orElseThrow(
+                            () -> new NotFoundException("The maintenance type id " + id +" was not found.")
+                    );
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public Iterable<MaintenanceTypeModel> findAllMaintenanceTypes() {
-        return repository.findAll();
+    public Iterable<MaintenanceTypeModel> findAllMaintenanceTypes() throws ExpiredJwtException{
+        try{
+            return repository.findAll();
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public MaintenanceTypeModel findMaintenanceTypeById(Long id) {
-        return findOrThrow(id);
+    public MaintenanceTypeModel findMaintenanceTypeById(Long id) throws ExpiredJwtException {
+        try{
+            if(!repository.existsById(id)) throw new NotFoundException("The maintenance type id " + id +" was not found.");
+            return findOrThrow(id);
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public void deleteMaintenanceTypeById(Long id) {
-        repository.deleteById(id);
+    public void deleteMaintenanceTypeById(Long id) throws ExpiredJwtException {
+        try{
+            if(!repository.existsById(id)) throw new NotFoundException("The maintenance type id " + id +" was not found.");
+            repository.deleteById(id);
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public MaintenanceTypeModel createMaintenanceType(MaintenanceTypeModel maintenanceType) {
-        return repository.save(maintenanceType);
+    public MaintenanceTypeModel createMaintenanceType(MaintenanceTypeModel maintenanceType) throws ExpiredJwtException {
+        try{
+            return repository.save(maintenanceType);
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public void updateMaintenanceType(Long id, MaintenanceTypeModel maintenanceType) {
-        findOrThrow(id);
-        repository.save(maintenanceType);
+    public void updateMaintenanceType(Long id, MaintenanceTypeModel maintenanceType) throws ExpiredJwtException {
+        try {
+            if(!repository.existsById(id)) throw new NotFoundException("The maintenance type id " + id +" was not found.");
+            findOrThrow(id);
+            repository.save(maintenanceType);
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 }
